Alteracoes nas exceptions, e funcionamento do endpoint de buscar todos os clientes

+++ src/main/java/com/transferencia/itau/json/ClienteResponse.java
@@ -12,6 +12,6 @@ public class ClienteResponse {
 
     private Long id;
     private String nome;
-    private Long numeroConta;
+    private double numeroConta;
     private Long saldo;
 }
